+++ aw_chromium/src/main/java/org/chromium/mojo/system/InvalidHandle.java
@@ -1,4 +1,4 @@
-// Copyright 2014 The Chromium Authors. All rights reserved.
+// Copyright 2014 The Chromium Authors
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
@@ -14,8 +14,7 @@ import java.util.List;
 /**
  * A handle that will always be invalid.
  */
-public class InvalidHandle implements UntypedHandle, MessagePipeHandle, ConsumerHandle,
-        ProducerHandle, SharedBufferHandle {
+public class InvalidHandle implements UntypedHandle, MessagePipeHandle, ConsumerHandle, ProducerHandle, SharedBufferHandle {
 
     /**
      * Instance singleton.
@@ -80,7 +79,7 @@ public class InvalidHandle implements UntypedHandle, MessagePipeHandle, Consumer
      * @see Handle#releaseNativeHandle()
      */
     @Override
-    public int releaseNativeHandle() {
+    public long releaseNativeHandle() {
         return 0;
     }
 
@@ -152,8 +151,7 @@ public class InvalidHandle implements UntypedHandle, MessagePipeHandle, Consumer
      * @see DataPipe.ProducerHandle#beginWriteData(int, DataPipe.WriteFlags)
      */
     @Override
-    public ByteBuffer beginWriteData(int numBytes,
-            DataPipe.WriteFlags flags) {
+    public ByteBuffer beginWriteData(int numBytes, DataPipe.WriteFlags flags) {
         throw new MojoException(MojoResult.INVALID_ARGUMENT);
     }
 
@@ -185,8 +183,7 @@ public class InvalidHandle implements UntypedHandle, MessagePipeHandle, Consumer
      * @see DataPipe.ConsumerHandle#beginReadData(int, DataPipe.ReadFlags)
      */
     @Override
-    public ByteBuffer beginReadData(int numBytes,
-            DataPipe.ReadFlags flags) {
+    public ByteBuffer beginReadData(int numBytes, DataPipe.ReadFlags flags) {
         throw new MojoException(MojoResult.INVALID_ARGUMENT);
     }
 
@@ -200,7 +197,7 @@ public class InvalidHandle implements UntypedHandle, MessagePipeHandle, Consumer
 
     /**
      * @see MessagePipeHandle#writeMessage(java.nio.ByteBuffer, java.util.List,
-     *      MessagePipeHandle.WriteFlags)
+     * MessagePipeHandle.WriteFlags)
      */
     @Override
     public void writeMessage(ByteBuffer bytes, List<? extends Handle> handles, WriteFlags flags) {
